@@ -1,26 +1,36 @@
 using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Net.Http;
+using System.Threading;
 using System.Threading.Tasks;
 using Http.Resilience.Internals;
-using Http.Resilience.Internals.Logging;
+using Http.Resilience.Logging;
+using Http.Resilience.Policies;
 
 namespace Http.Resilience
 {
     /// <summary>
-    /// HTTP retry helper which can be used to retry failed HTTP calls.
-    /// The retry behavior is configurable in <seealso cref="HttpRetryOptions"/>.
+    ///     HTTP retry helper which can be used to retry failed HTTP calls.
+    ///     The retry behavior is configurable in <seealso cref="HttpRetryOptions" />.
     /// </summary>
-    public class HttpRetryHelper
+    public class HttpRetryHelper : IHttpRetryHelper
     {
         private readonly string instance = Guid.NewGuid().ToString().Substring(0, 5).ToUpperInvariant();
-        private readonly HttpRetryOptions options;
 
-        private Func<Exception, bool> canRetryDelegate;
+        static readonly Lazy<IHttpRetryHelper> Implementation = new Lazy<IHttpRetryHelper>(CreateInstance, LazyThreadSafetyMode.PublicationOnly);
 
-        public HttpRetryOptions Options => this.options;
+        public static IHttpRetryHelper Current => Implementation.Value;
+
+        private static IHttpRetryHelper CreateInstance()
+        {
+            return new HttpRetryHelper();
+        }
+
+        private readonly IDictionary<Type, ICollection<IRetryPolicy>> retryPolicies = new Dictionary<Type, ICollection<IRetryPolicy>>();
 
         /// <summary>
-        /// Creates an instance of <seealso cref="HttpRetryHelper"/> with default <seealso cref="HttpRetryOptions"/>.
+        ///     Creates an instance of <seealso cref="HttpRetryHelper" /> with default <seealso cref="HttpRetryOptions" />.
         /// </summary>
         public HttpRetryHelper()
             : this(HttpRetryOptions.Default)
@@ -28,8 +38,8 @@ namespace Http.Resilience
         }
 
         /// <summary>
-        /// Creates an instance of <seealso cref="HttpRetryHelper"/> with default <seealso cref="HttpRetryOptions"/>
-        /// overriding <paramref name="maxRetries"/>.
+        ///     Creates an instance of <seealso cref="HttpRetryHelper" /> with default <seealso cref="HttpRetryOptions" />
+        ///     overriding <paramref name="maxRetries" />.
         /// </summary>
         public HttpRetryHelper(int maxRetries)
             : this(new HttpRetryOptions { MaxRetries = maxRetries })
@@ -37,56 +47,89 @@ namespace Http.Resilience
         }
 
         /// <summary>
-        /// Creates an instance of <seealso cref="HttpRetryHelper"/> with <paramref name="httpRetryOptions"/>.
+        ///     Creates an instance of <seealso cref="HttpRetryHelper" /> with <paramref name="httpRetryOptions" />.
         /// </summary>
         public HttpRetryHelper(HttpRetryOptions httpRetryOptions)
         {
-            this.options = httpRetryOptions;
+            this.Options = httpRetryOptions;
+            
+            this.AddOrUpdateRetryPolicy(new HttpMessageResponseRetryPolicy(this.Options));
+            this.AddOrUpdateRetryPolicy(new WebExceptionRetryPolicy(this.Options));
+            this.AddOrUpdateRetryPolicy(new SystemNetSocketExceptionRetryPolicy());
+            this.AddOrUpdateRetryPolicy(new SystemIOExceptionRetryPolicy());
+            this.AddOrUpdateRetryPolicy(new CurlExceptionRetryPolicy());
         }
 
+        public HttpRetryOptions Options { get; }
+
         /// <summary>
-        /// Calls <paramref name="action"/> synchronously.
+        ///     Calls <paramref name="action" /> synchronously.
         /// </summary>
-        public void Invoke(Action action)
+        public void Invoke(Action action, string actionName = nameof(Invoke))
         {
+            if (action == null)
+            {
+                throw new ArgumentNullException(nameof(action));
+            }
+
             AsyncHelper.RunSync(() => this.InvokeAsync(() =>
             {
                 action();
                 return Task.FromResult<object>(null);
-            }));
+            }, actionName));
         }
 
         /// <summary>
-        /// Calls <paramref name="function"/> synchronously and returns <typeparamref name="TResult"/>.
+        ///     Calls <paramref name="function" /> synchronously and returns <typeparamref name="TResult" />.
         /// </summary>
-        public TResult Invoke<TResult>(Func<TResult> function)
+        public TResult Invoke<TResult>(Func<TResult> function, string functionName = nameof(Invoke))
         {
+            if (function == null)
+            {
+                throw new ArgumentNullException(nameof(function));
+            }
+
             return AsyncHelper.RunSync(() => this.InvokeAsync(() =>
             {
                 return Task.FromResult(function());
-            }));
+            }, functionName));
         }
 
         /// <summary>
-        /// Calls <paramref name="function"/> asynchronously.
+        ///     Calls <paramref name="function" /> asynchronously.
         /// </summary>
-        public async Task InvokeAsync(Func<Task> function)
+        public async Task InvokeAsync(Func<Task> function, string functionName = nameof(InvokeAsync))
         {
+            if (function == null)
+            {
+                throw new ArgumentNullException(nameof(function));
+            }
+
             await this.InvokeAsync<object>(async () =>
             {
                 await function();
                 return Task.FromResult<object>(null);
-            });
+            }, functionName);
         }
 
         /// <summary>
-        /// Calls <paramref name="function"/> asynchronously and returns <typeparamref name="TResult"/>.
+        ///     Calls <paramref name="function" /> asynchronously and returns <typeparamref name="TResult" />.
         /// </summary>
-        public async Task<TResult> InvokeAsync<TResult>(Func<Task<TResult>> function)
+        public async Task<TResult> InvokeAsync<TResult>(Func<Task<TResult>> function,
+            string functionName = nameof(InvokeAsync))
         {
+            if (function == null)
+            {
+                throw new ArgumentNullException(nameof(function));
+            }
+
+            if (string.IsNullOrEmpty(functionName))
+            {
+                throw new ArgumentNullException(nameof(functionName));
+            }
+
             var currentAttempt = 1;
-            var maxAttempts = this.options.MaxRetries + 1;
-            var lastStatusCode = 0;
+            var maxAttempts = this.Options.MaxRetries + 1;
             var lastResult = default(TResult);
 
             while (true)
@@ -95,53 +138,152 @@ namespace Http.Resilience
 
                 try
                 {
-                    this.Log(LogLevel.Debug, $"Starting InvokeAsync... (Attempt {currentAttempt}/{maxAttempts})");
+                    this.Log(LogLevel.Info, $"Starting {functionName}... (Attempt {currentAttempt}/{maxAttempts})");
 
                     lastResult = await function();
                     if (lastResult is HttpResponseMessage httpResponseMessage)
                     {
-                        lastStatusCode = (int)httpResponseMessage.StatusCode;
-
-                        if (this.options.EnsureSuccessStatusCode)
-                        {
-                            httpResponseMessage.EnsureSuccessStatusCode();
-                        }
-                        else
-                        {
-                            if (HasRemainingAttempts(remainingAttempts) && 
-                                this.options.IsRetryableResponse(httpResponseMessage))
-                            {
-                                await this.SleepAsync(remainingAttempts);
-                                currentAttempt++;
-                                this.Log(LogLevel.Info, $"InvokeAsync --> Retry on HttpResponseMessage.StatusCode={(int)httpResponseMessage.StatusCode} ({httpResponseMessage.StatusCode})");
-                                continue;
-                            }
-                        }
+                        httpResponseMessage.EnsureSuccessStatusCode();
                     }
 
-                    this.Log(currentAttempt <= 1 ? LogLevel.Debug : LogLevel.Info, $"InvokeAsync finished successfully (Attempt {currentAttempt}/{maxAttempts})");
+                    this.Log(LogLevel.Info,
+                        $"{functionName} finished successfully (Attempt {currentAttempt}/{maxAttempts})");
                     return lastResult;
                 }
                 catch (Exception ex)
                 {
-                    this.Log(LogLevel.Error, $"InvokeAsync failed with exception (Attempt {currentAttempt}/{maxAttempts})");
+                    var hasRemainingAttempts = HasRemainingAttempts(remainingAttempts);
+
+                    this.Log(hasRemainingAttempts ? LogLevel.Debug : LogLevel.Error,
+                        $"{functionName} failed with exception (Attempt {currentAttempt}/{maxAttempts})");
 
                     var lastHttpResponseMessage = lastResult as HttpResponseMessage;
-                    if (HasRemainingAttempts(remainingAttempts) && 
-                        (NetworkHelper.IsTransientNetworkException(ex, lastHttpResponseMessage, this.options) || this.canRetryDelegate != null && this.canRetryDelegate(ex)))
+                    var retry = hasRemainingAttempts &&
+                                (this.EvaluateRetryPolicies(lastHttpResponseMessage) || this.EvaluateRetryPolicies(ex));
+                    if (retry)
                     {
                         await this.SleepAsync(remainingAttempts);
                         currentAttempt++;
-                        this.Log(LogLevel.Info, $"InvokeAsync --> Retry on {ex.GetType().Name}");
+                        this.Log(LogLevel.Info, $"{functionName} --> Retry on {ex.GetType().Name}");
                         continue;
+                    }
+
+                    if (lastHttpResponseMessage != null && !this.Options.EnsureSuccessStatusCode)
+                    {
+                        return lastResult;
                     }
 
                     throw;
                 }
             }
         }
+        
+        private bool EvaluateRetryPolicies(object parameter)
+        {
+            if (parameter == null)
+            {
+                return false;
+            }
+            
+            var paramType = parameter.GetType();
+            var applicableRetryPolicies = this.retryPolicies
+                .Where(p => p.Key.IsAssignableFrom(paramType))
+                .SelectMany(p => p.Value)
+                .ToList();
+            
+            foreach (var retryPolicy in applicableRetryPolicies)
+            {
+                var shouldRetry = retryPolicy.ShouldRetry(parameter);
+                this.Log(shouldRetry ? LogLevel.Info : LogLevel.Debug, $"{retryPolicy.GetType().Name}.ShouldRetry({paramType.Name}) returned {shouldRetry}");
+                if (shouldRetry)
+                {
+                    return true;
+                }
+            }
 
-        private static bool HasRemainingAttempts(int remainingAttempts) => remainingAttempts > 0;
+            return false;
+        }
+
+        public IHttpRetryHelper AddRetryPolicy<T>(IRetryPolicy<T> retryPolicy)
+        {
+            this.AddOrUpdateRetryPolicy(retryPolicy);
+            return this;
+        }
+
+        /// <summary>
+        ///     Custom retry decision logic if an unsuccessful http response is returned.
+        /// </summary>
+        public IHttpRetryHelper RetryOnHttpMessageResponse(Func<HttpResponseMessage, bool> handler)
+        {
+            if (handler == null)
+            {
+                throw new ArgumentNullException(nameof(handler));
+            }
+
+            this.AddOrUpdateRetryPolicy(new HttpMessageResponseRetryPolicyDelegate(handler));
+            return this;
+        }
+
+        /// <summary>
+        ///     Custom retry decision logic if an exception occurred.
+        /// </summary>
+        public IHttpRetryHelper RetryOnException(Func<Exception, bool> handler)
+        {
+            if (handler == null)
+            {
+                throw new ArgumentNullException(nameof(handler));
+            }
+
+            this.AddOrUpdateRetryPolicy(new ExceptionRetryPolicyDelegate(handler));
+            return this;
+        }
+
+        /// <summary>
+        ///     Custom retry decision logic if an exception of type <typeparamref name="TException" /> occurred.
+        /// </summary>
+        public IHttpRetryHelper RetryOnException<TException>(Func<TException, bool> handler) where TException : Exception
+        {
+            return this.RetryOnException(ex =>
+            {
+                if (ex is TException tex)
+                {
+                    return handler(tex);
+                }
+
+                return false;
+            });
+        }
+
+        private void AddOrUpdateRetryPolicy<T>(IRetryPolicy<T> retryPolicy)
+        {
+            if (retryPolicy == null)
+            {
+                throw new ArgumentNullException(nameof(retryPolicy));
+            }
+
+            if (this.retryPolicies.TryGetValue(typeof(T), out var policies))
+            {
+                var retryPolicyType = retryPolicy.GetType();
+                if (retryPolicyType != typeof(HttpMessageResponseRetryPolicyDelegate) && 
+                    retryPolicyType != typeof(ExceptionRetryPolicyDelegate) && 
+                    policies.Any(p => p.GetType() == retryPolicyType))
+                {
+                    throw new InvalidOperationException($"Retry policy of type {retryPolicyType.Name} is already added.");
+                }
+                
+                policies.Add(retryPolicy);
+            }
+            else
+            {
+                this.retryPolicies.Add(typeof(T), new List<IRetryPolicy> { retryPolicy });
+            }
+        }
+
+        private static bool HasRemainingAttempts(int remainingAttempts)
+        {
+            return remainingAttempts > 0;
+        }
+
         private static int CalculateRemainingAttempts(int currentAttempt, int maxAttempts)
         {
             return maxAttempts - currentAttempt;
@@ -159,37 +301,8 @@ namespace Http.Resilience
 
         protected virtual TimeSpan CalculateBackoff(int remainingAttempts)
         {
-            return BackoffTimerHelper.GetExponentialBackoff(this.options.MaxRetries - remainingAttempts + 1, this.options.MinBackoff, this.options.MaxBackoff, this.options.BackoffCoefficient);
-        }
-
-        /// <summary>
-        /// Custom retry decision logic if an exception occurred.
-        /// </summary>
-        public HttpRetryHelper RetryOnException(Func<Exception, bool> handler)
-        {
-            if (this.canRetryDelegate != null)
-            {
-                throw new InvalidOperationException($"{nameof(RetryOnException)} cannot be called more than once");
-            }
-
-            this.canRetryDelegate = handler;
-            return this;
-        }
-
-        /// <summary>
-        /// Custom retry decision logic if an exception of type <typeparamref name="TException"/> occurred.
-        /// </summary>
-        public HttpRetryHelper RetryOnException<TException>(Func<TException, bool> handler) where TException : Exception
-        {
-            return this.RetryOnException((ex) =>
-            {
-                if (ex is TException tex)
-                {
-                    return handler(tex);
-                }
-
-                return false;
-            });
+            return BackoffTimerHelper.GetExponentialBackoff(this.Options.MaxRetries - remainingAttempts + 1,
+                this.Options.MinBackoff, this.Options.MaxBackoff, this.Options.BackoffCoefficient);
         }
 
         private void Log(LogLevel logLevel, string message)
