@@ -4,20 +4,22 @@ using System.Net.Http;
 using System.Threading.Tasks;
 using Http.Resilience;
 using Http.Resilience.Internals;
+using Http.Resilience.Logging;
+using ResilienceConsole.Logging;
 
 namespace ResilienceConsole
 {
-    class Program
+    public static class Program
     {
         static async Task Main(string[] args)
         {
             Console.WriteLine("Http.Resilience Sample Console App");
             Console.WriteLine();
 
-            Http.Resilience.Internals.Logging.Logger.SetLogger(new ConsoleLogger());
+            Logger.SetLogger(new ConsoleLogger());
 
             // Example 1:
-            // Everything is okay. The first attempty succeeds
+            // Everything is okay. The first attempts succeeds
             // so that there is no further retry necessary.
             await Example1_OK();
 
